updates to AddEstimateLine.java

+++ app/src/main/java/com/example/metalconstructionsestimates/modules/estimateslines/AddEstimateLine.java
@@ -312,7 +312,7 @@ public class AddEstimateLine extends AppCompatActivity {
                         }
 
                         if (!Objects.requireNonNull(quantityTextInputEditText.getText()).toString().isEmpty()) {
-                            estimateLine.setQuantity(Integer.parseInt(quantityTextInputEditText.getText().toString()));
+                            estimateLine.setQuantity(Long.parseLong(quantityTextInputEditText.getText().toString()));
                         } else {
                             estimateLine.setQuantity(null);
                         }
